Extracted edgestore tests to own test file test_manager_edgestore
